SSH2Stream: fix terminal mode check in pty()
`modesToBytes()` tests `!RE_IS_NUM.test(opcode)`.  In the happy path, this
is always true: we looked up the name (key) in TERMINAL_MODE, and got
back an integer (it is a known terminal mode), or undefined (unknown).

We want to check `!RE_IS_NUM.test(key)` instead: if the key was a number,
then we looked up a number in TERMINAL_MODE, rather than a mode string.