SSH2Stream: Dispatch overloaded USERAUTH packet types (60-79)
The SSH protocol interprets packets with type 60 differently based on
the authentication method currently in progress (see RFC4252 and the
keyboard-interactive internet-draft):
  publickey: USERAUTH_PK_OK
  keyboard-interactive: USERAUTH_INFO_REQUEST (prompt)
  password: USERAUTH_PASSWORD_CHANGEREQ

Thus ssh2-streams needs to keep track of what authentication is in
progress when it is parsing a packet. Whenever a new authentication
method is attempted, that's the one that is in progress until some sort
of success or failure is received from the server (USERAUTH_SUCCESS,
USERAUTH_FAILURE, or USERAUTH_PK_OK). Keep the authentication methods
in progress in a queue (authsQueue) instead of a single slot
(authMethod).

Add test/test-packet60 with some basic packet tests and a client that
blindly sends a public key check, public key auth, password auth, and
keyboard interactive auth at once, exercising the client/server auth
queue.