SSH2Stream: ssh2-streams throws on receiving multiple KEXDH_INIT packet
ssh2-streams performs verification on receiving the packet that the
packet matches the expected packet. This check
however doesn't work for KEXDH_* packets as the context specific message codes
are not added to constants.Message object. As the packet types are missing in
constants.Message, packet verification in parsePacket doesn't work for KEXDH_* packets
subsequently creating an exception. Added code which verifies that the received
KEXDH_* packets is the next expected packet.

There was a second bug where after receiving NEWKEYS which causes
outstate.status to be OUT_READY, if KEXDH_* packet are received it also
created an exception. The check in parsePacket is skipped if the state is
OUT_READY and KEXDH_* packets are parsed and passed up the chain even though
this is an invalid packet. Added check to enusre if KEXDH_* packets are received
after OUT_READY session is disconnected.

Added unit tests which verifies the behavior of receiving
multiple KEXDH_INIT messages.

Server side code doesn't exist for dynamic group KEXDH algorithms hence
multiple KEXDH_GEX_* packets cannot be currently verified.